+++ .cargo/config.toml
@@ -15,4 +15,4 @@ build-std = ["std", "panic_abort"]
 [env]
 ESP_IDF_SDKCONFIG_DEFAULTS = ".github/configs/sdkconfig.defaults"
 ESP_IDF_VERSION = "v5.5.3"
-#MCU = "esp32c6"
+#MCU = "esp32h2"
