+++ src/wireless/wifi.rs
@@ -1,3 +1,4 @@
+#[cfg(esp_idf_bt_bluedroid_enabled)]
 use esp_idf_svc::bt::{self, BtDriver};
 use esp_idf_svc::eventloop::EspSystemEventLoop;
 use esp_idf_svc::hal::modem::Modem;
@@ -19,7 +20,9 @@ use crate::error::to_net_error;
 use crate::netif::{EspMatterNetStack, EspMatterNetif};
 use crate::wifi::EspMatterWifiCtl;
 
-use super::{EspWirelessMatterStack, GATTS_APP_ID};
+use super::EspWirelessMatterStack;
+#[cfg(esp_idf_bt_bluedroid_enabled)]
+use super::GATTS_APP_ID;
 
 /// A type alias for an ESP-IDF Matter stack running over Wifi (and BLE, during commissioning).
 pub type EspWifiMatterStack<'a, const B: usize, E> = EspWirelessMatterStack<'a, B, Wifi, E>;
@@ -101,10 +104,17 @@ impl Gatt for EspMatterWifi<'_, '_> {
     where
         A: GattTask,
     {
-        let bt = BtDriver::new(unsafe { self.modem.reborrow() }, Some(self.nvs.clone())).unwrap();
+        #[cfg(esp_idf_bt_bluedroid_enabled)]
+        let peripheral = {
+            let bt =
+                BtDriver::new(unsafe { self.modem.reborrow() }, Some(self.nvs.clone())).unwrap();
 
+            EspBtpGattPeripheral::<bt::Ble>::new(GATTS_APP_ID, bt, self.ble_context).unwrap()
+        };
+
+        #[cfg(not(esp_idf_bt_bluedroid_enabled))]
         let peripheral =
-            EspBtpGattPeripheral::<bt::Ble>::new(GATTS_APP_ID, bt, self.ble_context).unwrap();
+            EspBtpGattPeripheral::new(unsafe { self.modem.reborrow() }, self.ble_context).unwrap();
 
         task.run(peripheral).await
     }
@@ -171,10 +181,15 @@ impl WifiCoex for EspMatterWifi<'_, '_> {
 
         let net_ctl = EspMatterWifiCtl::new(wifi, self.sysloop.clone());
 
-        let bt = BtDriver::new(bt_p, Some(self.nvs.clone())).unwrap();
+        #[cfg(esp_idf_bt_bluedroid_enabled)]
+        let mut peripheral = {
+            let bt = BtDriver::new(bt_p, Some(self.nvs.clone())).unwrap();
 
-        let mut peripheral =
-            EspBtpGattPeripheral::<bt::Ble>::new(GATTS_APP_ID, bt, self.ble_context).unwrap();
+            EspBtpGattPeripheral::<bt::Ble>::new(GATTS_APP_ID, bt, self.ble_context).unwrap()
+        };
+
+        #[cfg(not(esp_idf_bt_bluedroid_enabled))]
+        let mut peripheral = EspBtpGattPeripheral::new(bt_p, self.ble_context).unwrap();
 
         task.run(
             EspMatterNetStack::new(),
